refactor: Session 상속 구조를 확장 -> 구성으로 변경
