+++ src/main/java/nextstep/courses/domain/Session.java
@@ -8,24 +8,25 @@ import nextstep.payments.domain.Payment;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
-public abstract class Session extends BaseEntity {
+public class Session extends BaseEntity {
 
     private final Long id;
     private final SessionType type;
     private final CoverImage coverImage;
     private final Period period;
     private Status status;
-    protected final Students students;
+    private final PaidCondition paidCondition;
+    private final Students students;
 
-    public Session(Long id, SessionType type, CoverImage coverImage, LocalDate startDate, LocalDate endDate) {
-        this(id, type, coverImage, new Period(startDate, endDate), Status.NOT_OPEN, new Students(), LocalDateTime.now(), null);
+    public static Session ofFree(Long id, CoverImage coverImage, LocalDate startDate, LocalDate endDate) {
+        return new Session(id, SessionType.FREE, coverImage, new Period(startDate, endDate), Status.NOT_OPEN, 0, 0L, LocalDateTime.now(), null);
     }
 
-    public Session(Long id, SessionType type, CoverImage coverImage, LocalDate startDate, LocalDate endDate, LocalDateTime createdAt, LocalDateTime updatedAt) {
-        this(id, type, coverImage, new Period(startDate, endDate), Status.NOT_OPEN, new Students(), createdAt, updatedAt);
+    public static Session ofPaid(Long id, CoverImage coverImage, LocalDate startDate, LocalDate endDate, int maxStudents, Long fee) {
+        return new Session(id, SessionType.PAID, coverImage, new Period(startDate, endDate), Status.NOT_OPEN, maxStudents, fee, LocalDateTime.now(), null);
     }
 
-    private Session(Long id, SessionType type, CoverImage coverImage, Period period, Status status, Students students, LocalDateTime createdAt, LocalDateTime updatedAt) {
+    private Session(Long id, SessionType type, CoverImage coverImage, Period period, Status status, int maxStudents, Long fee, LocalDateTime createdAt, LocalDateTime updatedAt) {
         super(createdAt, updatedAt);
         validateNotNull(id, coverImage, period);
         this.id = id;
@@ -33,7 +34,8 @@ public abstract class Session extends BaseEntity {
         this.coverImage = coverImage;
         this.period = period;
         this.status = status;
-        this.students = students;
+        this.paidCondition = new PaidCondition(maxStudents, fee);
+        this.students = new Students();
     }
 
     private void validateNotNull(Long id, CoverImage coverImage, Period period) {
@@ -42,7 +44,13 @@ public abstract class Session extends BaseEntity {
         }
     }
 
-    public abstract void register(Payment payment);
+    public void register(Payment payment) {
+        validateStatus();
+        if (type.isPaid()) {
+            paidCondition.validate(this.students, payment);
+        }
+        this.students.addStudent(payment.paidUser());
+    }
 
     protected void validateStatus() {
         if (!status.isOpen()) {
@@ -84,5 +92,4 @@ public abstract class Session extends BaseEntity {
     public LocalDate endDate() {
         return period.endDate();
     }
-
 }
