+++ src/test/java/nextstep/courses/domain/SessionTest.java
@@ -2,6 +2,8 @@ package nextstep.courses.domain;
 
 import nextstep.courses.exception.NotOpenSessionException;
 import nextstep.courses.exception.OutOfSessionException;
+import nextstep.courses.exception.OverMaxStudentsException;
+import nextstep.courses.exception.PaymentMismatchException;
 import nextstep.payments.domain.Payment;
 import nextstep.users.domain.NsUserTest;
 import org.junit.jupiter.api.DisplayName;
@@ -13,23 +15,48 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 class SessionTest {
 
-    private static final LocalDate START_DATE = LocalDate.of(2023, 12, 01);
-    private static final LocalDate END_DATE = LocalDate.of(2023, 12, 03);
+    private static final LocalDate START_DATE = LocalDate.of(2023, 12, 1);
+    private static final LocalDate END_DATE = LocalDate.of(2023, 12, 31);
 
     @Test
-    @DisplayName("수강 신청 시 강의 상태가 모집중이 아니면 예외를 던진다.")
+    @DisplayName("무료 강의 수강 신청 시 강의 상태가 모집중이 아니면 예외를 던진다.")
     void register_status_check() {
-        Session session = new FreeSession(1L, new CoverImage(1, "gif", 300, 200), START_DATE, END_DATE);
-        assertThatThrownBy(() -> session.register(Payment.ofFree(1L, NsUserTest.JAVAJIGI)))
+        Session freeSession = Session.ofFree(1L, coverImage(), START_DATE, END_DATE);
+        assertThatThrownBy(() -> freeSession.register(Payment.ofFree(1L, NsUserTest.JAVAJIGI)))
                 .isInstanceOf(NotOpenSessionException.class);
     }
 
     @Test
     @DisplayName("강의 상태를 모집중으로 변경 시 현재 날짜가 강의 기간에 속하지 않으면 예외를 던진다.")
     void register_open() {
-        Session session = new FreeSession(1L, new CoverImage(1, "gif", 300, 200), START_DATE, END_DATE);
+        Session session = Session.ofFree(1L, coverImage(), START_DATE, LocalDate.of(2023, 12, 2));
         assertThatThrownBy(() -> session.openSession())
                 .isInstanceOf(OutOfSessionException.class);
     }
 
+    @Test
+    @DisplayName("유료 강의 수강 신청 시 최대 수강 인원을 초과하면 예외를 던진다.")
+    void register_over_students() {
+        Session paidSession = Session.ofPaid(1L, coverImage(), START_DATE, END_DATE, 1, 10_000L);
+        paidSession.openSession();
+
+        paidSession.register(Payment.ofPaid(1L, 1L, NsUserTest.JAVAJIGI, 10_000L));
+
+        assertThatThrownBy(() -> paidSession.register(Payment.ofPaid(2L, 1L, NsUserTest.SANJIGI, 8_000L)))
+                .isInstanceOf(OverMaxStudentsException.class);
+    }
+
+    @Test
+    @DisplayName("유료 강의 수강 신청 시 결제금액과 수강료가 일치하는지 확인하지 않으면 예외를 던진다.")
+    void session_fee_test() {
+        Session paidSession = Session.ofPaid(1L, coverImage(), START_DATE, END_DATE, 1, 10_000L);
+        paidSession.openSession();
+        
+        assertThatThrownBy(() -> paidSession.register(Payment.ofPaid(2L, 1L, NsUserTest.SANJIGI, 8_000L)))
+                .isInstanceOf(PaymentMismatchException.class);
+    }
+
+    private static CoverImage coverImage() {
+        return new CoverImage(1, "gif", 300, 200);
+    }
 }
